requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
PyMuPDF>=1.22.5
python-docx>=0.8.11
azure-ai-inference==1.0.0b9
xlsxwriter>=3.1.0
toml>=0.10.2